% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-plot.R
\name{plot.lt_matrix}
\alias{plot.lt_matrix}
\title{Plot a La Terra matrix}
\usage{
\method{plot}{lt_matrix}(x, ...)
}
\arguments{
\item{x}{An \code{lt_matrix}.}

\item{...}{Reserved for a future plot specification.}
}
\value{
No value; always errors in this skeleton release.
}
\description{
Plot semantics are intentionally not frozen in LT001-A1A.
}
