% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-relative.R
\name{ratio_state}
\alias{ratio_state}
\alias{unavailability_reason}
\alias{exact_zero}
\alias{ratio_availability}
\alias{positive_log_availability}
\alias{positive_log_values}
\title{Materialize log-relative state and component accessors}
\usage{
ratio_state(x)

unavailability_reason(x)

exact_zero(x)

ratio_availability(x)

positive_log_availability(x)

positive_log_values(x)
}
\arguments{
\item{x}{An \code{lt_log_relative} object.}
}
\value{
An aligned matrix.
}
\description{
Materialize log-relative state and component accessors
}
