% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-spec.R
\name{validate_lt_analysis_spec}
\alias{validate_lt_analysis_spec}
\title{Validate a La Terra analysis specification}
\usage{
validate_lt_analysis_spec(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a La Terra analysis specification
}
