% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{validate_lt_association}
\alias{validate_lt_association}
\title{Validate an association point-estimate result}
\usage{
validate_lt_association(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate an association point-estimate result
}
