% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{validate_lt_association_domain}
\alias{validate_lt_association_domain}
\title{Validate a Layer-5 association domain}
\usage{
validate_lt_association_domain(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a Layer-5 association domain
}
