% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-relative-association.R
\name{validate_lt_boundary_association}
\alias{validate_lt_boundary_association}
\title{Validate a boundary association}
\usage{
validate_lt_boundary_association(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a boundary association
}
