% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-relative-association.R
\name{validate_lt_boundary_calibration}
\alias{validate_lt_boundary_calibration}
\title{Validate a boundary calibration bundle}
\usage{
validate_lt_boundary_calibration(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a boundary calibration bundle
}
