% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-relative-association.R
\name{validate_lt_boundary_component_calibration}
\alias{validate_lt_boundary_component_calibration}
\title{Validate one boundary-component calibration}
\usage{
validate_lt_boundary_component_calibration(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate one boundary-component calibration
}
