% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_state.R
\name{validate_lt_branch_state}
\alias{validate_lt_branch_state}
\title{Validate a branch-associated state object}
\usage{
validate_lt_branch_state(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a branch-associated state object
}
