% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{validate_lt_domain}
\alias{validate_lt_domain}
\title{Validate a La Terra taxon domain}
\usage{
validate_lt_domain(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a La Terra taxon domain
}
