% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-relative.R
\name{validate_lt_log_relative}
\alias{validate_lt_log_relative}
\title{Validate a boundary-aware log-relative object}
\usage{
validate_lt_log_relative(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a boundary-aware log-relative object
}
