% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{validate_lt_matrix}
\alias{validate_lt_matrix}
\title{Validate a La Terra matrix}
\usage{
validate_lt_matrix(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly, or an error describing the violated invariant.
}
\description{
Validate a La Terra matrix
}
