% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{validate_lt_run}
\alias{validate_lt_run}
\title{Validate an audited La Terra run container}
\usage{
validate_lt_run(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate an audited La Terra run container
}
