% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait.R
\name{validate_lt_trait}
\alias{validate_lt_trait}
\title{Validate a La Terra trait}
\usage{
validate_lt_trait(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a La Terra trait
}
